requirements.txt
configparser~=5.2.0
#logging
python-gitlab~=3.2.0
requests~=2.27.1